Install: don't overwrite systemd config if it exists

+++ scripts/install.sh
@@ -80,20 +80,21 @@ trap install_success EXIT
 # Everything from this point onwards is optional.
 
 configure_systemd() {
-    if ! id ollama >/dev/null 2>&1; then
-        status "Creating ollama user..."
-        $SUDO useradd -r -s /bin/false -m -d /usr/share/ollama ollama
-    fi
-    if getent group render >/dev/null 2>&1; then
-        status "Adding ollama user to render group..."
-        $SUDO usermod -a -G render ollama
-    fi
+    if [ ! -f /etc/systemd/system/ollama.service ]; then
+        if ! id ollama >/dev/null 2>&1; then
+            status "Creating ollama user..."
+            $SUDO useradd -r -s /bin/false -m -d /usr/share/ollama ollama
+        fi
+        if getent group render >/dev/null 2>&1; then
+            status "Adding ollama user to render group..."
+            $SUDO usermod -a -G render ollama
+        fi
 
-    status "Adding current user to ollama group..."
-    $SUDO usermod -a -G ollama $(whoami)
+        status "Adding current user to ollama group..."
+        $SUDO usermod -a -G ollama $(whoami)
 
-    status "Creating ollama systemd service..."
-    cat <<EOF | $SUDO tee /etc/systemd/system/ollama.service >/dev/null
+        status "Creating ollama systemd service..."
+        cat <<EOF | $SUDO tee /etc/systemd/system/ollama.service >/dev/null
 [Unit]
 Description=Ollama Service
 After=network-online.target
@@ -109,6 +110,8 @@ Environment="PATH=$PATH"
 [Install]
 WantedBy=default.target
 EOF
+    fi
+
     SYSTEMCTL_RUNNING="$(systemctl is-system-running || true)"
     case $SYSTEMCTL_RUNNING in
         running|degraded)
